requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-filter==2.3.0
django-storages==1.10
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.91
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
